feat: Update Terraform provider source in main.tf
The `source` attribute in the `required_providers` block of `main.tf` has been updated to point to the new location of the Terraform provider. This change ensures that the correct provider is used for the project.

Note: This commit message assumes that the purpose of the code changes is to update the Terraform provider source. If there are other changes in the code that are not mentioned in the provided diff, please provide more context so that a more accurate commit message can be generated.