feat: Add data source for retrieving datacenters
This commit adds a new data source function `dataSourceDatacenter` to retrieve datacenters. It includes optional parameters for filtering by `id`, `name`, and `country`. The function reads the datacenters from an external service and matches the provided parameters to find a matching datacenter. If a match is found, the datacenter details are set in the resource data. If no match is found, an error is returned with a list of available datacenters.

Note: This commit message assumes that the purpose of the code changes is to add a data source for retrieving datacenters. If there are other changes in the code that are not mentioned in the provided diff, please provide more context so that a more accurate commit message can be generated.