feat: Update goreleaser version and arguments
The goreleaser version has been updated to '~> v2' and the release arguments have been changed to '--clean'. This update ensures that the code is using the latest version of goreleaser and includes the necessary arguments for a clean release.

Note: This commit message assumes that the purpose of the code changes is to update the goreleaser version and arguments. If there are other changes in the code that are not mentioned in the provided diff, please provide more context so that a more accurate commit message can be generated.